block, bfq: save & resume weight on a queue merge/split
bfq saves the state of a queue each time a merge occurs, to be
able to resume such a state when the queue is associated again
with its original process, on a split.

Unfortunately bfq does not save & restore also the weight of the
queue. If the weight is not correctly resumed when the queue is
recycled, then the weight of the recycled queue could differ
from the weight of the original queue.

This commit adds the missing save & resume of the weight.

Tested-by: Holger Hoffstätte <holger@applied-asynchrony.com>
Tested-by: Oleksandr Natalenko <oleksandr@natalenko.name>
Signed-off-by: Francesco Pollicino <fra.fra.800@gmail.com>
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit fffca087d587b03d0d0dca2e86bf8e688fbf2c18)
Signed-off-by: Vaibhav Rustagi <vaibhavrustagi@google.com>

BUG=b:143356772
TEST=cos_tryjob
SOURCE=UPSTREAM(fffca087d587),CHROMEOS(ec4762f621a1)

Change-Id: I55f9d9aaeb7b0560edf40eafb81bb168c39fc2e9