kbuild: Make NOSTDINC_FLAGS a simply expanded variable
During a simple no-op (nothing changed) build I saw 39 invocations of
the C compiler with the argument -print-file-name=include.  We dont
need to call the C compiler 39 times for this--one time will suffice.

Lets change NOSTDINC_FLAGS to a simply expanded variable to avoid
this since there doesnt appear to be any reason it should be
recursively expanded.

On my build this shaved ~400 ms off my no-op build.

Note that the recursive expansion seems to date back to the (really
old) commit e8f5bdb02ce0 ([PATCH] Makefile include path ordering).
Its a little unclear to me if the point of that patch was to switch
the variable to be recursively expanded (which it did) or to avoid
directly assigning to NOSTDINC_FLAGS (AKA to switch to +=) because
someone else (out of tree?) was setting it.  I presume later since if
the only goal was to switch to recursive expansion the patch would
have just removed the :.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
(cherry picked from commit 0c22be0712b8771011880972aa30d3412aa3334e)
Signed-off-by: Vaibhav Rustagi <vaibhavrustagi@google.com>

BUG=b:143356772
TEST=cos_tryjob
SOURCE=UPSTREAM(0c22be0712b8),CHROMEOS(8d2c0bea489c)

Change-Id: Ib13024ae3d27dd06272bfde18f1ae9b5c27f54de