ACPI: button: reinitialize button state upon resume
With commit dfa46c50f65b (ACPI / button: Fix an issue in
button.lid_init_state=ignore mode), the lid device is considered to be
not compliant to SW_LID if the Lid state is unchanged when updating it.

This is not wrong, but we overlooked the resume case, where Lid state is
updated unconditionally in the button driver .resume() callback. And this
results in warning message ACPI: button: The lid device is not compliant
to  SW_LID. after resume, if the machine is suspended with Lid opened and
then resumed with Lid opened.

Fix this by flushing the cached lid state before updating the Lid device
in .resume() callback.

Fixes: dfa46c50f65b (ACPI / button: Fix an issue in button.lid_init_state=ignore mode)
Reported-and-tested-by: Zhao Lijian <lijian.zhao@intel.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 13e962140be671f31a011543f11477af67a6c33e)
Signed-off-by: Vaibhav Rustagi <vaibhavrustagi@google.com>

BUG=b:143356772
TEST=cos_tryjob
SOURCE=UPSTREAM(13e962140be6),CHROMEOS(30035fdbc25b)

Change-Id: I561485149ad8782e2a35c81c8d4e71254e46487b