block, bfq: fix rq_in_driver check in bfq_update_inject_limit
One of the cases where the parameters for injection may be updated is
when there are no more in-flight I/O requests. The number of in-flight
requests is stored in the field bfqd->rq_in_driver of the descriptor
bfqd of the device. So, the controlled condition is
bfqd->rq_in_driver == 0.

Unfortunately, this is wrong because, the instruction that checks this
condition is in the code path that handles the completion of a
request, and, in particular, the instruction is executed before
bfqd->rq_in_driver is decremented in such a code path.

This commit fixes this issue by just replacing 0 with 1 in the
comparison.

Reported-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Tested-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit db599f9ed9bd31b018b6c48ad7c6b21d5b790ecf)
Signed-off-by: Vaibhav Rustagi <vaibhavrustagi@google.com>

BUG=b:143356772
TEST=cos_tryjob
SOURCE=UPSTREAM(db599f9ed9bd),CHROMEOS(07be472ee5d3)

Change-Id: Ic728ab00292f6b1c7625ec49bc2bbf8deb3f713d