block, bfq: update base request service times when possible
I/O injection gets reduced if it increases the request service times
of the victim queue beyond a certain threshold.  The threshold, in its
turn, is computed as a function of the base service time enjoyed by
the queue when it undergoes no injection.

As a consequence, for injection to work properly, the above base value
has to be accurate. In this respect, such a value may vary over
time. For example, it varies if the size or the spatial locality of
the I/O requests in the queue change. It is then important to update
this value whenever possible. This commit performs this update.

Reported-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Tested-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit 24792ad01cb659c8b5899de2af6e8ca250f93df3)
Signed-off-by: Vaibhav Rustagi <vaibhavrustagi@google.com>

BUG=b:143356772
TEST=cos_tryjob
SOURCE=UPSTREAM(24792ad01cb6),CHROMEOS(2a6ea5a2252e)

Change-Id: I16563bcd538a8f86455a7f9f8185ae795e2f49c7