CHROMIUM: config: Use CONFIG_HZ_250 for arm32
Lower the rate of the system tick from 1000 Hz to 250 Hz for
arm32/veyron devices. This improves the scores of some benchmarks.
The 3.14 kernel uses a tick rate of 100 Hz, a higher tick rate is
supposed to improve responsiveness, use 250 Hz as compromise.

The config is the result of executing the following commands on
top of a normalized configuration:

  echo CONFIG_HZ_250=y >> chromeos/config/armel/common.config

  ./chromeos/scripts/kernelconfig olddefconfig

BUG=chromium:989730
TEST=Octane and webgl.Aquarium scores improve for veyron

Change-Id: I544281c2a14c848bfb18d6fdf0958876fc99aadb
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1729844
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 36f504838a7d8eafcc80d5134368c69efa1debd0)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1746754