gve: Add Gvnic stats AQ command and ethtool show/set-priv-flags.
Changes:
- Add a new flag in service_task_flags. Check both this flag and
ethtool flag when handle report stats. Update the stats when user turns
ethtool flag on.

- In order to expose the NIC stats to the guest even when the ethtool flag
is off, share the address and length of report at setup. When the
ethtool flag turned off, zero off the gve stats instead of detaching the
report. Only detach the report in free_stats_report.

- Adds the NIC stats to ethtool stats. These stats are always
exposed to guest no matter the report stats flag is turned
on or off.

- Update gve stats once every 20 seconds.

- Add a field for the interval of updating stats report to the AQ
command. It will be exposed to USPS so that USPS can use the same
interval to update its stats in the report.

Patches squashed into this patch:
- Add Gvnic stats AQ command and ethtool show/set-priv-flags.
- Fix handle_status and handle_report_stats for report_stats flags.
- Fix exposing cross-reporting stats to always share stats report.
- Exposes NIC stats to the guest by 'ethtool -S'.
- Fix attached/detached the stats report.
- Fix the stats report to be attached after configure_device_resources and
detached before deconfigure_device_resources.
- Adds a timer to update the gve stats in stats report.
- Fix the timer to use jiffies instead of msecs.
- Fix c90 compiler warning about vla
- Add 'interval' field to report stats AQ command.

BUG=b/154868226

Change-Id: Ie67818eade47fc4769ac2c208a442cfb71e4a852
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/1582
Reviewed-by: Yangchun Fu <yangchun@google.com>
Reviewed-by: Roy Yang <royyang@google.com>