LAKITU: security: bug fix: create exit event only when thread group dead
The previous hook on task exit was waiting for all tasks to be gone but
the kernel is very asynchronous when it is exiting threads. This
change moves the hook to be called only when the whole thread group is
dead.

BUG=b:143630036
TEST=Ran all internal testing.

Signed-off-by: Thomas Garnier <thgarnie@chromium.org>
Change-Id: Ifd997567fc69258901f125c5281e3d0b17a85271
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1973031
Reviewed-by: Vaibhav Rustagi <vaibhavrustagi@google.com>
Reviewed-by: Roy Yang <royyang@google.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
(cherry picked from commit fcc231356fa2904d7bfa308d2464581f298768a7)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1979122