ext4: add prefetch_block_bitmaps mount options
For file systems where we can afford to keep the buddy bitmaps cached,
we can speed up initial writes to large file systems by starting to
load the block allocation bitmaps as soon as the file system is
mounted.  This won't work well for _super_ large file systems, or
memory constrained systems, so we only enable this when it is
requested via a mount option.

Addresses-Google-Bug: 159488342
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iad3356da63f624f5dc6345874c4ec9dda6dff51d
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/3303
Reviewed-by: Theodore Tso <tytso@google.com>