arch/x86: allow to execute syscalls via process_vm_exec
process_vm_exec allows to execute code in an address space of another
process. It changes the current address space to the target address
space and resume the current process with registers from sigcontex that
is passed in the arguments.

This changes adds the PROCESS_VM_EXEC_SYSCALL flag and if it is set
process_vm_exec will execute a system call with arguments from sigcontext.

process_vm_exec retuns 0 if the system call has been executed and an error
code in other cases.

A return code of the system call can be found in a proper register in
sigcontext.

BUG=183648601
TEST=Compiled and verified locally
SOURCE=FROMLIST(https://lore.kernel.org/linux-api/20210414055217.543246-1-avagin@gmail.com/`)
RELEASE_NOTE=Port process_vm_exec to the COS kernel.

Signed-off-by: Andrei Vagin <avagin@google.com>
Change-Id: Ia26939a75ebbfde0943242d0b324a90c38603a11
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/17934
Reviewed-by: Oleksandr Tymoshenko <ovt@google.com>
Tested-by: Oleksandr Tymoshenko <ovt@google.com>