arch/arm64: implement the process_vm_exec syscall
This change introduces the new system call:
process_vm_exec(pid_t pid, struct process_vm_exec_context *uctx,
	    unsigned long flags, siginfo_t * uinfo,
	    sigset_t *sigmask, size_t sizemask)

process_vm_exec allows to execute the current process in an address
space of another process.

process_vm_exec swaps the current address space with an address space of a
specified process, sets a state from uctx and resumes the process.  When a
process receives a signal or calls a system call, process_vm_exec saves the
process state back to uctx, restores the origin address space, restores the
origin process state, and returns to userspace.

If it was interrupted by a signal and the signal is in the user_mask,
the signal is dequeued and information about it is saved in uinfo.
If process_vm_exec is interrupted by a system call, a synthetic siginfo
for the SIGSYS signal is generated.

The behavior of this system call is similar to PTRACE_SYSEMU but
everything is happing in the context of one process, so
process_vm_exec shows a better performance.

PTRACE_SYSEMU is primarily used to implement sandboxes (application
kernels) like User-mode Linux or gVisor. These type of sandboxes
intercepts applications system calls and acts as the guest kernel.
A simple benchmark, where a "tracee" process executes systems calls in a
loop and a "tracer" process traps syscalls and handles them just
incrementing the tracee instruction pointer to skip the syscall
instruction shows that process_vm_exec works more than 5 times faster
than PTRACE_SYSEMU.

BUG=183648601
TEST=Compiled and verified locally
SOURCE=FROMLIST(https://lore.kernel.org/linux-api/20210414055217.543246-1-avagin@gmail.com/`)
RELEASE_NOTE=Port process_vm_exec to the COS kernel.

Signed-off-by: Andrei Vagin <avagin@google.com>
Change-Id: I703f8b1836d62f9b25592d89d872c5b9c44e4b46
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/17935
Reviewed-by: Oleksandr Tymoshenko <ovt@google.com>
Tested-by: Oleksandr Tymoshenko <ovt@google.com>