LAKITU: Add stub IDs for some BPF helpers that exist upstream
BPF helpers are identified by an ID which is defined in enum
bpf_func_id via this X-macro trickery. In a subsequent commit I'll
backport a new BPF helper. Because of the enum approach to the helper
IDs, if we did this naïvely the ID given to the backported helper
would not match the one used upstream, which means BPF programs
compiled against upstream headers would not work.

As a simple solution to this, add the IDs of the helpers in between
ima_inode_hash and the helper I'll backport (snprintf).

This doesn't change the runtime code of the kernel; if someone tries
to load a program referring to these helpers it will be rejected with
an unknown function call, just the same as if this commit was never
applied.

If you're rebasing this commit onto v5.13+ you can just drop it.

The commits that actually add these helpers upstream are:

4f19cab76136 ("bpf: Add a bpf_sock_from_file helper")
34b2021cc616 ("bpf: bpf_fib_lookup return MTU value as output when
              looked up")
69c087ba6225 ("bpf: Change return value of verifier function
              add_subprog()")

Signed-off-by: Brendan Jackman <jackmanb@google.com>

BUG=b:217368493

Change-Id: I39db1bdf3d402b18409376de7479673aa9bbf793