x86/mm: Validate memory when changing the C-bit
Add the needed functionality to change pages state from shared
to private and vice-versa using the Page State Change VMGEXIT as
documented in the GHCB spec.

[acdunlap: The previous version of this patch assumes that
sev_es_ghcb_hv_call takes 6 arguments, but this version of the code is
missing the patch 007faec014cb5 that adds an extra argument.  The extra
argument 'bool set_ghcb_msr' only changes the code if it is false, and
in the previous patch it is set to true, so it's OK to remove it.]

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lore.kernel.org/r/20220307213356.2797205-22-brijesh.singh@amd.com
Change-Id: Ie239ad58e954f712b38b34dcd609dffb373aafd4