KVM: x86: Move lookup of indexed CPUID leafs to helper
Determining which CPUID leafs have significant ECX/index values is
also needed by guest kernel code when doing SEV-SNP-validated CPUID
lookups. Move this to common code to keep future updates in sync.

Modified from the original verison of this commit so
cpuid_function_is_indexed returns false for the cases 0x1d and 0x1e.
This is because the patch 690a757d610e kvm: x86: Add CPUID support for
Intel AMX is not in this repository.

Signed-off-by: Michael Roth <michael.roth@amd.com>
Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Venu Busireddy <venu.busireddy@oracle.com>
Link: https://lore.kernel.org/r/20220307213356.2797205-31-brijesh.singh@amd.com
Change-Id: I79ca239cf6ac8a3741d908c60c5a62613ab4407b