x86/sev: Provide support for SNP guest request NAEs
Version 2 of GHCB specification provides SNP_GUEST_REQUEST and
SNP_EXT_GUEST_REQUEST NAE that can be used by the SNP guest to
communicate with the PSP.

While at it, add a snp_issue_guest_request() helper that will be used by
driver or other subsystem to issue the request to PSP.

See SEV-SNP firmware and GHCB spec for more details.

[acdunlap: The previous version of this patch assumes that
sev_es_ghcb_hv_call takes 6 arguments, but this version of the code is
missing the patch 007faec014cb5 that adds an extra argument.  The extra
argument 'bool set_ghcb_msr' only changes the code if it is false, and
in the previous patch it is set to true, so it's OK to remove it.]

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lore.kernel.org/r/20220307213356.2797205-42-brijesh.singh@amd.com
Change-Id: I1f47a92425f97d14b9e7b01d644ea5c9947e62cb