dma-buf: add option text so it can be enabled by users
This reverts commit 3b32a592ea6e4 ("dma-buf: drop option text so users
don't select it.") to allow users select DMA_SHARED_BUFFER option
explicitly. This is required for out-of-tree modules that depend on
dma-buf functionality, for instance drivers that rely on Google's Gasket
framework.

On general-purpose Linux systems the option is usually implicitely
enabled in the default kernel config by subsystems that use it but
systems with smaller kernel footprint have to rely on side-effects
from enabling unrelated options to get dma-buf functionality built.

BUG=b/214598029
TEST=presubmit
RELEASE_NOTE=None

Signed-off-by: Oleksandr Tymoshenko <ovt@google.com>
Change-Id: I369d3d6a1c706c7d5207f8a7a9c22a467fbc5acd
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/32940
Reviewed-by: Roy Yang <royyang@google.com>
Reviewed-by: Meena Shanmugam <meenashanmugam@google.com>
Tested-by: Cusky Presubmit Bot <presubmit@cos-infra-prod.iam.gserviceaccount.com>