x86/mm: Remove extra argument to sev_es_ghcb_hv_call
The previous patch x86/mm: Validate memory when changing the C-bit
assumes that sev_es_ghcb_hv_call takes 6 arguments, but this version of
the code is missing the patch 007faec014cb5 that adds an extra argument.
The extra argument 'bool set_ghcb_msr' only changes the code if it is
false, and in the previous patch it is set to true, so it's OK to remove
it.

Change-Id: Icabcfb92e5ea7f803d85fd0ca8a6f69b5c45117a