x86/kvm: Revert previous commit's changes to kvm.c
In this patch set we are trying to not modify kvm. The previous patch
was necessary for some changes it made guest-side, but it also made
changes to kvm.c that we do not want and do not compile without extra
patches. This commit reverts the changes to kvm.c.

Change-Id: Ic7e6320af984bfff56b94b07ae2b70825653fc66