kernel: Add kernel config fragment that is required to run xfstests.
Currently kernel configuration from https://github.com/tytso/xfstests-bld is used to build the kernel to run xfstests. The reason for using the kernel config from xfstests-bld repo is, there are certain
configurations needed to be enabled or disabled to use the cos kernel
for xfstesting. There is a gap between the configurations that customers
use and we are testing. In order to minimize the gap, the required
configurations for xfstests are identified and added as a separate
kernel configuration fragment.

BUG=b/240174731
TEST=presubmit,validation
RELEASE_NOTE=None

Change-Id: I0f4611f0eaa018d41faa33d1abe92b91519d2552
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/37409
Reviewed-by: Oleksandr Tymoshenko <ovt@google.com>
Tested-by: Cusky Presubmit Bot <presubmit@cos-infra-prod.iam.gserviceaccount.com>